app-server/requirements.txt
viam-sdk
fastapi
pylint
black
uvicorn
python-dotenv